pyproject.toml
[project]
name = "agentic-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "fastparquet>=2026.3.0",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "langsmith>=0.7.25",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "langchain[openai]>=1.2.13",
    "langchain-cloudflare>=0.2.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-deepseek>=1.0.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "langgraph-supervisor>=0.0.31",
    "lmstudio>=1.5.0",
    "pypdf>=6.7.5",
    "python-terrier>=1.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "txtai>=9.5.0",
    "langchain-anthropic>=1.4.0",
    "langchain-experimental>=0.4.1",
    "matplotlib>=3.10.8",
    "langchain-text-splitters>=1.1.1",
    "bs4>=0.0.2",
    "langchain-mistralai>=1.1.2",
    "chromadb>=1.5.5",
    "backoff>=2.2.1",
]
[tool.uv.sources]
chunking-evaluation = { git = "https://github.com/brandonstarxel/chunking_evaluation.git" }
